+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Improvements
 - Arch
   - not rechecking sub-dependencies of an AUR dependency to be installed
+  - allowing AUR packages to be installed as dependencies of a repository package 
 
 ### Fixes
 - General
@@ -17,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Arch
   - not upgrading a package when a dependent package relies on a specific version with epoch (e.g: alsa-plugins 1:1.2.6-1 would not be upgraded to 1:1.2.6-2 because lib32-alsa-plugins relies on 1:1.2.6)
   - not informing all the provided packages on the transaction context to the dependency sorting algorithm (could lead to a wrong installation order)
+  - not displaying all possible AUR providers for a given dependency
 
 ### UI
     - new logo by [DN-debug](https://github.com/DN-debug)